module-1/pyproject.toml
[project]
name = "module-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.29.0",
    "python-dotenv>=1.1.1",
]
